Update configuration and improve chat statistical analysis
This update changes the configuration from .ini to .py files for better integration. It also refactors and enhances the chat statistics functionality: more precise and additional statistics are calculated, including averages per day and symbols per message. Now, the statistics are calculated not just for a single chat, but for multiple chats defined in a list. Pyrogram has been updated to version 2.0.106 for stability improvements.